requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-dotenv==1.0.1
langchain==0.2.11
langchain-community==0.2.10
langchain-groq==0.1.5
